Move realtime clock code into realtime.c
